add sealed blob type
